extract versions typescript code
