+++ src/components/common/dev/Versions.vue
@@ -22,13 +22,9 @@
 </template>
 
 <script lang="ts">
-    import { useStore } from "@/use/store/store";
-    import AuthenticationManagement from "@/api/AuthenticationManagement";
-    import CourseManagement from "@/api/CourseManagement";
-    import UserManagement from "@/api/UserManagement";
     import { ref, onBeforeMount } from "vue";
     import LoadingSpinner from "@/components/common/loading/Spinner.vue";
-    import MatriculationManagement from "@/api/MatriculationManagement";
+    import { version, getVersions } from "@/use/helpers/Versions";
 
     export default {
         components: {
@@ -39,71 +35,16 @@
         setup(props: any, { emit }: any) {
             let busy = ref(false);
 
-            interface version {
-                name: String;
-                version: String;
-                link: String;
-            }
+            let versions = ref([] as version[]);
 
-            let versions: version[] = [];
-            onBeforeMount(() => {
-                getVersions();
+            onBeforeMount(async () => {
+                await getBackendVersions();
             });
 
-            async function getVersions() {
+            async function getBackendVersions() {
                 busy.value = true;
-                let store = useStore();
-
-                let frontEndVersion = "v" + process.env.VUE_APP_VERSION;
-
-                let authenticationManagementVersion = await AuthenticationManagement.getVersion();
-                let courseManagementVersion = await CourseManagement.getVersion();
-                let userManagementVersion = await UserManagement.getVersion();
-                let matriculationManagementVersion = await MatriculationManagement.getVersion();
-
-                versions.push({
-                    name: "Frontend",
-                    version: frontEndVersion,
-                    link: "https://github.com/upb-uc4/ui-web/blob/" + frontEndVersion + "/CHANGELOG.md",
-                });
-                versions.push({
-                    name: "Authentication Management",
-                    version: authenticationManagementVersion,
-                    link:
-                        "https://github.com/upb-uc4/University-Credits-4.0/blob/" +
-                        "authentication-" +
-                        authenticationManagementVersion +
-                        "/product_code/authentication_service/CHANGELOG.md",
-                });
-                versions.push({
-                    name: "Course Management",
-                    version: courseManagementVersion,
-                    link:
-                        "https://github.com/upb-uc4/University-Credits-4.0/blob/" +
-                        "course-" +
-                        courseManagementVersion +
-                        "/product_code/course_service/CHANGELOG.md",
-                });
-                versions.push({
-                    name: "User Management",
-                    version: userManagementVersion,
-                    link:
-                        "https://github.com/upb-uc4/University-Credits-4.0/blob/" +
-                        "user-" +
-                        userManagementVersion +
-                        "/product_code/user_service/CHANGELOG.md",
-                });
-                versions.push({
-                    name: "Matriculation Management",
-                    version: matriculationManagementVersion,
-                    link:
-                        "https://github.com/upb-uc4/University-Credits-4.0/blob/" +
-                        "matriculation-" +
-                        matriculationManagementVersion +
-                        "/product_code/matriculation_service/CHANGELOG.md",
-                });
-
-                emit("versions", versions);
+                versions.value = await getVersions();
+                emit("versions", versions.value);
                 busy.value = false;
             }
